create common InfoRow component
