+++ ui/src/components/person/PersonCard.tsx
@@ -1,7 +1,8 @@
-import { Badge, Card, Group, Stack, Text } from "@mantine/core";
+import { Badge, Card, Stack } from "@mantine/core";
 import type { PersonOutput } from "bridg-ts";
 import dayjs from "dayjs";
 import { useTranslation } from "react-i18next";
+import { InfoRow } from "../InfoRow";
 
 interface PersonCardProps {
   person: PersonOutput;
@@ -60,23 +61,3 @@ export const PersonCard = ({ person }: PersonCardProps) => {
     </Card>
   );
 };
-
-interface InfoRowProps {
-  label: string;
-  value?: React.ReactNode;
-  children?: React.ReactNode;
-}
-
-function InfoRow({ label, value, children }: InfoRowProps) {
-  return (
-    <Group>
-      <Text fw={600} w={150} c="dimmed">
-        {label}:
-      </Text>
-      <Text>
-        {value || "N/A"}
-        {children}
-      </Text>
-    </Group>
-  );
-}
